refactor: extract IO utilities to _output.py
