+++ src/con_duct/duct_main.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python3
 from __future__ import annotations
-from collections.abc import Iterable
 from dataclasses import asdict
 from importlib.metadata import version
 import json
@@ -11,7 +10,6 @@ import shutil
 import signal
 import socket
 import subprocess
-import sys
 import threading
 import time
 from typing import IO, Any, Optional, TextIO
@@ -25,6 +23,7 @@ from con_duct._models import (
     SessionMode,
     SystemInfo,
 )
+from con_duct._output import TailPipe, prepare_outputs, remove_files, safe_close_files
 from con_duct._sampling import _get_sample
 from con_duct._signals import SigIntHandler
 
@@ -280,105 +279,6 @@ def monitor_process(
         if stop_event.wait(timeout=sample_interval):
             lgr.debug("Breaking out because stop event was set")
             break
-
-
-class TailPipe:
-    """TailPipe simultaneously streams to an output stream (stdout or stderr) and a specified file."""
-
-    TAIL_CYCLE_TIME = 0.01
-
-    def __init__(self, file_path: str, buffer: IO[bytes]) -> None:
-        self.file_path = file_path
-        self.buffer = buffer
-        self.stop_event: threading.Event | None = None
-        self.infile: IO[bytes] | None = None
-        self.thread: threading.Thread | None = None
-
-    def start(self) -> None:
-        self.stop_event = threading.Event()
-        self.infile = open(self.file_path, "rb")
-        self.thread = threading.Thread(target=self._tail, daemon=True)
-        self.thread.start()
-
-    def fileno(self) -> int:
-        assert self.infile is not None
-        return self.infile.fileno()
-
-    def _catch_up(self) -> None:
-        assert self.infile is not None
-        data = self.infile.read()
-        if data:
-            self.buffer.write(data)
-            self.buffer.flush()
-
-    def _tail(self) -> None:
-        assert self.stop_event is not None
-        try:
-            while not self.stop_event.is_set():
-                self._catch_up()
-                time.sleep(TailPipe.TAIL_CYCLE_TIME)
-            # After stop event, collect and passthrough data one last time
-            self._catch_up()
-        except Exception:
-            raise
-        finally:
-            self.buffer.flush()
-
-    def close(self) -> None:
-        assert self.stop_event is not None
-        assert self.thread is not None
-        assert self.infile is not None
-        self.stop_event.set()
-        self.thread.join()
-        self.infile.close()
-
-
-def prepare_outputs(
-    capture_outputs: Outputs,
-    outputs: Outputs,
-    log_paths: LogPaths,
-) -> tuple[TextIO | TailPipe | int | None, TextIO | TailPipe | int | None]:
-    stdout: TextIO | TailPipe | int | None
-    stderr: TextIO | TailPipe | int | None
-
-    if capture_outputs.has_stdout():
-        if outputs.has_stdout():
-            stdout = TailPipe(log_paths.stdout, buffer=sys.stdout.buffer)
-            stdout.start()
-        else:
-            stdout = open(log_paths.stdout, "w")
-    elif outputs.has_stdout():
-        stdout = None
-    else:
-        stdout = subprocess.DEVNULL
-
-    if capture_outputs.has_stderr():
-        if outputs.has_stderr():
-            stderr = TailPipe(log_paths.stderr, buffer=sys.stderr.buffer)
-            stderr.start()
-        else:
-            stderr = open(log_paths.stderr, "w")
-    elif outputs.has_stderr():
-        stderr = None
-    else:
-        stderr = subprocess.DEVNULL
-    return stdout, stderr
-
-
-def safe_close_files(file_list: Iterable[Any]) -> None:
-    for f in file_list:
-        try:
-            f.close()
-        except Exception:
-            pass
-
-
-def remove_files(log_paths: LogPaths, assert_empty: bool = False) -> None:
-    for _, file_path in log_paths:
-        if os.path.exists(file_path):
-            if assert_empty:
-                assert os.stat(file_path).st_size == 0
-            os.remove(file_path)
 
 
 def execute(
