+++ test/duct_main/test_prepare_outputs.py
@@ -3,14 +3,14 @@ import subprocess
 from unittest.mock import MagicMock, call, patch
 from utils import MockStream
 from con_duct._models import LogPaths, Outputs
-from con_duct.duct_main import prepare_outputs
+from con_duct._output import prepare_outputs
 
 
 @patch("builtins.open", new_callable=MagicMock)
-@patch("con_duct.duct_main.TailPipe")
-@patch("con_duct.duct_main.LogPaths")
-@patch("con_duct.duct_main.sys.stderr", new_callable=MockStream)
-@patch("con_duct.duct_main.sys.stdout", new_callable=MockStream)
+@patch("con_duct._output.TailPipe")
+@patch("con_duct._output.LogPaths")
+@patch("con_duct._output.sys.stderr", new_callable=MockStream)
+@patch("con_duct._output.sys.stdout", new_callable=MockStream)
 def test_prepare_outputs_capture_none_output_none(
     _mock_stdout: MockStream,
     _mock_stderr: MockStream,
@@ -28,10 +28,10 @@ def test_prepare_outputs_capture_none_output_none(
 
 
 @patch("builtins.open", new_callable=MagicMock)
-@patch("con_duct.duct_main.TailPipe")
-@patch("con_duct.duct_main.LogPaths")
-@patch("con_duct.duct_main.sys.stderr", new_callable=MockStream)
-@patch("con_duct.duct_main.sys.stdout", new_callable=MockStream)
+@patch("con_duct._output.TailPipe")
+@patch("con_duct._output.LogPaths")
+@patch("con_duct._output.sys.stderr", new_callable=MockStream)
+@patch("con_duct._output.sys.stdout", new_callable=MockStream)
 def test_prepare_outputs_capture_none_output_stdout(
     _mock_stdout: MockStream,
     _mock_stderr: MockStream,
@@ -49,10 +49,10 @@ def test_prepare_outputs_capture_none_output_stdout(
 
 
 @patch("builtins.open", new_callable=MagicMock)
-@patch("con_duct.duct_main.TailPipe")
-@patch("con_duct.duct_main.LogPaths")
-@patch("con_duct.duct_main.sys.stderr", new_callable=MockStream)
-@patch("con_duct.duct_main.sys.stdout", new_callable=MockStream)
+@patch("con_duct._output.TailPipe")
+@patch("con_duct._output.LogPaths")
+@patch("con_duct._output.sys.stderr", new_callable=MockStream)
+@patch("con_duct._output.sys.stdout", new_callable=MockStream)
 def test_prepare_outputs_capture_none_output_stderr(
     _mock_stdout: MockStream,
     _mock_stderr: MockStream,
@@ -70,10 +70,10 @@ def test_prepare_outputs_capture_none_output_stderr(
 
 
 @patch("builtins.open", new_callable=MagicMock)
-@patch("con_duct.duct_main.TailPipe")
-@patch("con_duct.duct_main.LogPaths")
-@patch("con_duct.duct_main.sys.stderr", new_callable=MockStream)
-@patch("con_duct.duct_main.sys.stdout", new_callable=MockStream)
+@patch("con_duct._output.TailPipe")
+@patch("con_duct._output.LogPaths")
+@patch("con_duct._output.sys.stderr", new_callable=MockStream)
+@patch("con_duct._output.sys.stdout", new_callable=MockStream)
 def test_prepare_outputs_capture_none_output_all(
     _mock_stdout: MockStream,
     _mock_stderr: MockStream,
@@ -91,10 +91,10 @@ def test_prepare_outputs_capture_none_output_all(
 
 
 @patch("builtins.open", new_callable=MagicMock)
-@patch("con_duct.duct_main.TailPipe")
-@patch("con_duct.duct_main.LogPaths")
-@patch("con_duct.duct_main.sys.stderr", new_callable=MockStream)
-@patch("con_duct.duct_main.sys.stdout", new_callable=MockStream)
+@patch("con_duct._output.TailPipe")
+@patch("con_duct._output.LogPaths")
+@patch("con_duct._output.sys.stderr", new_callable=MockStream)
+@patch("con_duct._output.sys.stdout", new_callable=MockStream)
 def test_prepare_outputs_capture_stdout_output_none(
     _mock_stdout: MockStream,
     _mock_stderr: MockStream,
@@ -112,10 +112,10 @@ def test_prepare_outputs_capture_stdout_output_none(
 
 
 @patch("builtins.open", new_callable=MagicMock)
-@patch("con_duct.duct_main.TailPipe")
-@patch("con_duct.duct_main.LogPaths")
-@patch("con_duct.duct_main.sys.stderr", new_callable=MockStream)
-@patch("con_duct.duct_main.sys.stdout", new_callable=MockStream)
+@patch("con_duct._output.TailPipe")
+@patch("con_duct._output.LogPaths")
+@patch("con_duct._output.sys.stderr", new_callable=MockStream)
+@patch("con_duct._output.sys.stdout", new_callable=MockStream)
 def test_prepare_outputs_capture_stdout_output_stdout(
     mock_stdout: MockStream,
     _mock_stderr: MockStream,
@@ -135,10 +135,10 @@ def test_prepare_outputs_capture_stdout_output_stdout(
 
 
 @patch("builtins.open", new_callable=MagicMock)
-@patch("con_duct.duct_main.TailPipe")
-@patch("con_duct.duct_main.LogPaths")
-@patch("con_duct.duct_main.sys.stderr", new_callable=MockStream)
-@patch("con_duct.duct_main.sys.stdout", new_callable=MockStream)
+@patch("con_duct._output.TailPipe")
+@patch("con_duct._output.LogPaths")
+@patch("con_duct._output.sys.stderr", new_callable=MockStream)
+@patch("con_duct._output.sys.stdout", new_callable=MockStream)
 def test_prepare_outputs_capture_stdout_output_stderr(
     _mock_stdout: MockStream,
     _mock_stderr: MockStream,
@@ -156,10 +156,10 @@ def test_prepare_outputs_capture_stdout_output_stderr(
 
 
 @patch("builtins.open", new_callable=MagicMock)
-@patch("con_duct.duct_main.TailPipe")
-@patch("con_duct.duct_main.LogPaths")
-@patch("con_duct.duct_main.sys.stderr", new_callable=MockStream)
-@patch("con_duct.duct_main.sys.stdout", new_callable=MockStream)
+@patch("con_duct._output.TailPipe")
+@patch("con_duct._output.LogPaths")
+@patch("con_duct._output.sys.stderr", new_callable=MockStream)
+@patch("con_duct._output.sys.stdout", new_callable=MockStream)
 def test_prepare_outputs_capture_stdout_output_all(
     mock_stdout: MockStream,
     _mock_stderr: MockStream,
@@ -179,10 +179,10 @@ def test_prepare_outputs_capture_stdout_output_all(
 
 
 @patch("builtins.open", new_callable=MagicMock)
-@patch("con_duct.duct_main.TailPipe")
-@patch("con_duct.duct_main.LogPaths")
-@patch("con_duct.duct_main.sys.stderr", new_callable=MockStream)
-@patch("con_duct.duct_main.sys.stdout", new_callable=MockStream)
+@patch("con_duct._output.TailPipe")
+@patch("con_duct._output.LogPaths")
+@patch("con_duct._output.sys.stderr", new_callable=MockStream)
+@patch("con_duct._output.sys.stdout", new_callable=MockStream)
 def test_prepare_outputs_capture_stderr_output_none(
     _mock_stdout: MockStream,
     _mock_stderr: MockStream,
@@ -200,10 +200,10 @@ def test_prepare_outputs_capture_stderr_output_none(
 
 
 @patch("builtins.open", new_callable=MagicMock)
-@patch("con_duct.duct_main.TailPipe")
-@patch("con_duct.duct_main.LogPaths")
-@patch("con_duct.duct_main.sys.stderr", new_callable=MockStream)
-@patch("con_duct.duct_main.sys.stdout", new_callable=MockStream)
+@patch("con_duct._output.TailPipe")
+@patch("con_duct._output.LogPaths")
+@patch("con_duct._output.sys.stderr", new_callable=MockStream)
+@patch("con_duct._output.sys.stdout", new_callable=MockStream)
 def test_prepare_outputs_capture_stderr_output_stdout(
     _mock_stdout: MockStream,
     _mock_stderr: MockStream,
@@ -221,10 +221,10 @@ def test_prepare_outputs_capture_stderr_output_stdout(
 
 
 @patch("builtins.open", new_callable=MagicMock)
-@patch("con_duct.duct_main.TailPipe")
-@patch("con_duct.duct_main.LogPaths")
-@patch("con_duct.duct_main.sys.stderr", new_callable=MockStream)
-@patch("con_duct.duct_main.sys.stdout", new_callable=MockStream)
+@patch("con_duct._output.TailPipe")
+@patch("con_duct._output.LogPaths")
+@patch("con_duct._output.sys.stderr", new_callable=MockStream)
+@patch("con_duct._output.sys.stdout", new_callable=MockStream)
 def test_prepare_outputs_capture_stderr_output_stderr(
     _mock_stdout: MockStream,
     mock_stderr: MockStream,
@@ -244,10 +244,10 @@ def test_prepare_outputs_capture_stderr_output_stderr(
 
 
 @patch("builtins.open", new_callable=MagicMock)
-@patch("con_duct.duct_main.TailPipe")
-@patch("con_duct.duct_main.LogPaths")
-@patch("con_duct.duct_main.sys.stderr", new_callable=MockStream)
-@patch("con_duct.duct_main.sys.stdout", new_callable=MockStream)
+@patch("con_duct._output.TailPipe")
+@patch("con_duct._output.LogPaths")
+@patch("con_duct._output.sys.stderr", new_callable=MockStream)
+@patch("con_duct._output.sys.stdout", new_callable=MockStream)
 def test_prepare_outputs_capture_stderr_output_all(
     _mock_stdout: MockStream,
     mock_stderr: MockStream,
@@ -267,10 +267,10 @@ def test_prepare_outputs_capture_stderr_output_all(
 
 
 @patch("builtins.open", new_callable=MagicMock)
-@patch("con_duct.duct_main.TailPipe")
-@patch("con_duct.duct_main.LogPaths")
-@patch("con_duct.duct_main.sys.stderr", new_callable=MockStream)
-@patch("con_duct.duct_main.sys.stdout", new_callable=MockStream)
+@patch("con_duct._output.TailPipe")
+@patch("con_duct._output.LogPaths")
+@patch("con_duct._output.sys.stderr", new_callable=MockStream)
+@patch("con_duct._output.sys.stdout", new_callable=MockStream)
 def test_prepare_outputs_capture_all_output_none(
     _mock_stdout: MockStream,
     _mock_stderr: MockStream,
@@ -290,10 +290,10 @@ def test_prepare_outputs_capture_all_output_none(
 
 
 @patch("builtins.open", new_callable=MagicMock)
-@patch("con_duct.duct_main.TailPipe")
-@patch("con_duct.duct_main.LogPaths")
-@patch("con_duct.duct_main.sys.stderr", new_callable=MockStream)
-@patch("con_duct.duct_main.sys.stdout", new_callable=MockStream)
+@patch("con_duct._output.TailPipe")
+@patch("con_duct._output.LogPaths")
+@patch("con_duct._output.sys.stderr", new_callable=MockStream)
+@patch("con_duct._output.sys.stdout", new_callable=MockStream)
 def test_prepare_outputs_capture_all_output_stdout(
     mock_stdout: MockStream,
     _mock_stderr: MockStream,
@@ -311,10 +311,10 @@ def test_prepare_outputs_capture_all_output_stdout(
 
 
 @patch("builtins.open", new_callable=MagicMock)
-@patch("con_duct.duct_main.TailPipe")
-@patch("con_duct.duct_main.LogPaths")
-@patch("con_duct.duct_main.sys.stderr", new_callable=MockStream)
-@patch("con_duct.duct_main.sys.stdout", new_callable=MockStream)
+@patch("con_duct._output.TailPipe")
+@patch("con_duct._output.LogPaths")
+@patch("con_duct._output.sys.stderr", new_callable=MockStream)
+@patch("con_duct._output.sys.stdout", new_callable=MockStream)
 def test_prepare_outputs_capture_all_output_stderr(
     _mock_stdout: MockStream,
     mock_stderr: MockStream,
@@ -334,10 +334,10 @@ def test_prepare_outputs_capture_all_output_stderr(
 
 
 @patch("builtins.open", new_callable=MagicMock)
-@patch("con_duct.duct_main.TailPipe")
-@patch("con_duct.duct_main.LogPaths")
-@patch("con_duct.duct_main.sys.stderr", new_callable=MockStream)
-@patch("con_duct.duct_main.sys.stdout", new_callable=MockStream)
+@patch("con_duct._output.TailPipe")
+@patch("con_duct._output.LogPaths")
+@patch("con_duct._output.sys.stderr", new_callable=MockStream)
+@patch("con_duct._output.sys.stdout", new_callable=MockStream)
 def test_prepare_outputs_capture_all_output_all(
     mock_stdout: MockStream,
     mock_stderr: MockStream,
